removed about page and added toolbox page
