+++ src/app/(pages)/(home)/page.tsx
@@ -3,7 +3,7 @@ import Content from "./sections/Content";
 import Intro from "./sections/Intro";
 import Knowledge from "./sections/Knowledge";
 import LogoBanner from "./sections/LogoBanner";
-import Products from "./sections/Products";
+
 
 
 const HomePage = () => (
@@ -11,7 +11,6 @@ const HomePage = () => (
 		<Intro />
 		<Knowledge />
 		<LogoBanner/>
-		<Products/>
 		<Content />
 		<CTA />
 	</>
