+++ src/app/(pages)/(home)/sections/Knowledge.tsx
@@ -36,11 +36,12 @@ const Knowledge = () => {
 						Move faster with our cutting-edge AI-powered tools
 					</h2>
 					<p className="mt-2 text-3xl font-bold tracking-tight text-white sm:text-4xl">
-					Everyone desires extra time, and our ChatBot provides everything you need to make that a reality.
+					Everyone desires extra time, <br/> we make that a reality.
 					</p>
 					<p className="mt-6 text-lg leading-8 text-gray-300">
 						Get the answers you need <b>fast</b>.
 					</p>
+					<br/>
 					<p>
 						At <b>INGRAM</b>, we believe that the power of AI should be
 						accessible to all, without compromising security. Discover how
