+++ src/components/Navbar.tsx
@@ -51,10 +51,10 @@ const Navbar = () => {
 						</li>
 						<li>
 							<Link
-								href="/about"
+								href="/product"
 								className="block py-2 pl-3 pr-4 text-white rounded hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-red-500 md:p-0"
 							>
-								About
+								ToolBox
 							</Link>
 						</li>
 						<li>
